Move Azure login code to EE
